Update README.md
My goal with this is to create a list of road blocks and check them off as they get resolved.

Day 1. I am at this time still hopeful that I will learn to program, and set up vs code to work for me. Though the set back today was hard to see.